Video-Summarizer/requirements.txt
ipykernel
moviepy
openai-whisper 
transformers 
pytube 
flask 
setuptools 

-e .